MadeChangesToPowerups

+++ Asteroids V2/Assets/_Scripts/GameController.cs
@@ -11,7 +11,7 @@ public class GameController : MonoBehaviour {
 	public int hazardCount;
 	public float spawnWait;
 	public float startWait;
-	public float wavecount;
+	public int wavecount;
 
 	public Text scoretext;
 	int score;
@@ -26,38 +26,40 @@ public class GameController : MonoBehaviour {
 		StartCoroutine (SpawnWaves ());
 
 	}
-	
+
 	// Update is called once per frame
-	void Update () {
-		//increases difficulty over time
-		if (wavecount >= 5) {
-			spawnWait = .8f;
-		}
-		if (wavecount >= 10) {
-			spawnWait = .7f;
-		}
-		if (wavecount >= 15) {
-			spawnWait = .6f;
-		}
-		if (wavecount >= 25) {
-			spawnWait = .5f;
-		}
-		if (wavecount >= 35) {
-			spawnWait = .4f;
-		}
-		if (wavecount >= 50) {
-			spawnWait = .3f;
-		}
-		if (wavecount >= 70) {
-			spawnWait = .2f;
-		}
-		if (wavecount >= 100) {
-			spawnWait = .1f;
-		}
-		if (wavecount >= 150) {
-			spawnWait = .05f;
-		}
+	void Update () 
+	{
+			//increases difficulty over time
+			if (wavecount >= 5) {
+				spawnWait = .8f;
+			}
+			if (wavecount >= 10) {
+				spawnWait = .7f;
+			}
+			if (wavecount >= 15) {
+				spawnWait = .6f;
+			}
+			if (wavecount >= 25) {
+				spawnWait = .5f;
+			}
+			if (wavecount >= 35) {
+				spawnWait = .4f;
+			}
+			if (wavecount >= 50) {
+				spawnWait = .3f;
+			}
+			if (wavecount >= 70) {
+				spawnWait = .2f;
+			}
+			if (wavecount >= 100) {
+				spawnWait = .1f;
+			}
+			if (wavecount >= 150) {
+				spawnWait = .05f;
+			}
 	}
+
 
 	IEnumerator SpawnWaves()
 	{
@@ -115,5 +117,5 @@ public class GameController : MonoBehaviour {
 		Quaternion spawnRotation = Quaternion.identity;
 		Instantiate (hazard, spawnRight, spawnRotation);
 	}
-
+	
 }
